+++ spec/rubocop/cop/rails/refute_methods_spec.rb
@@ -1,53 +1,122 @@
 # frozen_string_literal: true
 
-RSpec.describe RuboCop::Cop::Rails::RefuteMethods do
-  subject(:cop) { described_class.new }
+RSpec.describe RuboCop::Cop::Rails::RefuteMethods, :config do
+  subject(:cop) { described_class.new(config) }
 
-  it 'registers an offense and correct using `refute` with a single argument' do
-    expect_offense(<<~RUBY)
-      refute foo
-      ^^^^^^ Prefer `assert_not` over `refute`.
-    RUBY
-
-    expect_correction(<<~RUBY)
-      assert_not foo
-    RUBY
+  let(:config) do
+    RuboCop::Config.new('Rails/RefuteMethods' => cop_config)
   end
 
-  it 'registers an offense and corrects using `refute` ' \
-    'with multiple arguments' do
-    expect_offense(<<~RUBY)
-      refute foo, bar, baz
-      ^^^^^^ Prefer `assert_not` over `refute`.
-    RUBY
-
-    expect_correction(<<~RUBY)
-      assert_not foo, bar, baz
-    RUBY
+  let(:cop_config) do
+    {
+      'EnforcedStyle' => enforced_style,
+      'SupportedStyles' => %w[assert_not refute]
+    }
   end
 
-  it 'registers an offense when using `refute_empty`' do
-    expect_offense(<<~RUBY)
-      refute_empty foo
-      ^^^^^^^^^^^^ Prefer `assert_not_empty` over `refute_empty`.
-    RUBY
+  context 'when EnforcedStyle is `assert_not`' do
+    let(:enforced_style) { 'assert_not' }
 
-    expect_correction(<<~RUBY)
-      assert_not_empty foo
-    RUBY
+    it 'registers an offense and correct using `refute` ' \
+       'with a single argument' do
+      expect_offense(<<~RUBY)
+        refute foo
+        ^^^^^^ Prefer `assert_not` over `refute`.
+      RUBY
+
+      expect_correction(<<~RUBY)
+        assert_not foo
+      RUBY
+    end
+
+    it 'registers an offense and corrects using `refute` ' \
+       'with multiple arguments' do
+      expect_offense(<<~RUBY)
+        refute foo, bar, baz
+        ^^^^^^ Prefer `assert_not` over `refute`.
+      RUBY
+
+      expect_correction(<<~RUBY)
+        assert_not foo, bar, baz
+      RUBY
+    end
+
+    it 'registers an offense when using `refute_empty`' do
+      expect_offense(<<~RUBY)
+        refute_empty foo
+        ^^^^^^^^^^^^ Prefer `assert_not_empty` over `refute_empty`.
+      RUBY
+
+      expect_correction(<<~RUBY)
+        assert_not_empty foo
+      RUBY
+    end
+
+    it 'does not registers an offense when using `assert_not` ' \
+       'with a single argument' do
+      expect_no_offenses(<<~RUBY)
+        assert_not foo
+      RUBY
+    end
+
+    it 'does not registers an offense when using `assert_not` ' \
+       'with a multiple arguments' do
+      expect_no_offenses(<<~RUBY)
+        assert_not foo, bar, baz
+      RUBY
+    end
   end
 
-  it 'does not registers an offense when using `assert_not` ' \
-     'with a single argument' do
-    expect_no_offenses(<<~RUBY)
-      assert_not foo
-    RUBY
-  end
+  context 'when EnforcedStyle is `refute`' do
+    let(:enforced_style) { 'refute' }
 
-  it 'does not registers an offense when using `assert_not` ' \
-     'with a multiple arguments' do
-    expect_no_offenses(<<~RUBY)
-      assert_not foo, bar, baz
-    RUBY
+    it 'registers an offense and correct using `assert_not` ' \
+       'with a single argument' do
+      expect_offense(<<~RUBY)
+        assert_not foo
+        ^^^^^^^^^^ Prefer `refute` over `assert_not`.
+      RUBY
+
+      expect_correction(<<~RUBY)
+        refute foo
+      RUBY
+    end
+
+    it 'registers an offense and corrects using `assert_not` ' \
+       'with multiple arguments' do
+      expect_offense(<<~RUBY)
+        assert_not foo, bar, baz
+        ^^^^^^^^^^ Prefer `refute` over `assert_not`.
+      RUBY
+
+      expect_correction(<<~RUBY)
+        refute foo, bar, baz
+      RUBY
+    end
+
+    it 'registers an offense when using `assert_not_empty`' do
+      expect_offense(<<~RUBY)
+        assert_not_empty foo
+        ^^^^^^^^^^^^^^^^ Prefer `refute_empty` over `assert_not_empty`.
+      RUBY
+
+      expect_correction(<<~RUBY)
+        refute_empty foo
+      RUBY
+    end
+
+    it 'does not registers an offense when using `refute` ' \
+       'with a single argument' do
+      expect_no_offenses(<<~RUBY)
+        refute foo
+      RUBY
+    end
+
+    it 'does not registers an offense when using `refute` ' \
+       'with a multiple arguments' do
+      expect_no_offenses(<<~RUBY)
+        refute foo, bar, baz
+      RUBY
+    end
   end
 end